requirements.txt
# Core
python-dotenv>=1.0.1
pyyaml>=6.0.2
pandas>=2.2.2
numpy>=1.26.4
scipy>=1.13.1
requests>=2.32.3
pytz>=2024.1
tzlocal>=5.2
APScheduler>=3.10.4
loguru>=0.7.2
tenacity>=8.5.0
dataclasses-json>=0.6.7

# Trading / Indicators / Backtesting
TA-Lib>=0.4.28; platform_system != "Windows" or platform_machine != "x86" # install requires system lib; see README for Windows wheels
pandas_ta>=0.3.14b0
backtrader>=1.9.78.123
yfinance>=0.2.43

# IG API
trading-ig>=0.0.25 # https://github.com/ig-python/trading-ig
lightstreamer-client>=1.0.3
# Optional helper library
ig-markets-rest-api-python-library>=0.0.8

# Monitoring dashboard
streamlit>=1.39.0

# Testing
pytest>=8.3.3
pytest-cov>=5.0.0

psycopg2-binary>=2.9.9
